Update sentora_install.sh
The EPEL download code was removed because base RHEL repos come preloaded with the most recent version of EPEL already defined. So the system will download an install the most compatible version of the library for the system that runs the installer.

Note: If this step fails, the installer will not install proftd, the mysql extension or other required software, causing several issues. I've had this happen on at least 5 stock CentOS systems, and was able to rectify this issue by using this installer but changing the same code.